Auto-README/requirements.txt
# requirements.txt
google-generativeai>=0.3.0
streamlit>=1.10.0
python-dotenv>=0.19.0
gitpython>=3.1.0